Fix re-open bug.
When dismiss dialog, mainScope.cancel() is called.
Once this cancel is called, this scope seems no more usable.
So recreate scope when cancel. Also, move channel listener for startListening and call it whenever dialog re-appear.
I don't know what is the opposite of dismiss, so I use onPrepare for this purpose.